Breed in UI (#9)
* trigger breeding from UI

* make duration an int

* elitism count and tournament size from UI of ProblemSetup

* also multiple steps in a background thread

* corrected test error

* added scale

* control scale with a slider

* more in background

* use the population recursively

* improved UI

* corrections from MR remarks

+++ src/main/java/cj/software/genetics/schedule/util/WorkerService.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Service
 public class WorkerService {
     @SuppressWarnings("squid:S3518")    // zero division is not possible
-    public double calcDuration(Worker worker) {
-        double result = 0.0;
+    public int calcDuration(Worker worker) {
+        int result = 0;
         List<Task> tasks = worker.getTasks();
         if (tasks != null && !tasks.isEmpty()) {
             for (Task task : tasks) {
